Fix a problem with fractional seconds.
As specified in #1866, the `updated_at` column is used sometimes by
removing the `usec` of the time, and sometimes not. In recent versions
of MySQL, it actually saves the fractional portion, which means that
Paperclip will generate different hashes sometimes. This commit is
basically the workaround from that Issue, as a stopgap until a better
solution is found.